music_backend/requirements.txt
Django
textblob
djangorestframework
tensorflow
scikit-learn
django-environ
opencv-python
numpy
requests
Pillow
django-cors-headers
psycopg2
djangorestframework